log: add operator<<-based formatters for scoped enums and streamable classes
Adds two constrained std::formatter specializations inside the
MAVSDK_HAS_STD_FORMAT guard in log.h:

1. Scoped enum formatter: matches any `enum class` that has an
   `operator<<` (i.e. all MAVSDK-generated Result/Mode/Status/etc.
   enums), so they print their named value (e.g. "Success") instead
   of a raw integer.

2. Streamable class formatter: matches any class/struct with
   `operator<<` that isn't already handled by std (string,
   string_view, filesystem::path), so ParamValue and similar types
   format naturally.

With these in place, remove the now-unnecessary `static_cast<int>()`
casts on MAVSDK scoped enums and `.get_string()` calls on ParamValue
that were added as a workaround across mavsdk_server service impls,
system tests, camera_impl, mavlink_parameter_{client,server,subscription},
and mavlink_component_metadata.

Legitimate casts on external C enums (CURLcode, MAV_TYPE,
COMP_METADATA_TYPE, MAV_CMD, MAV_MISSION_RESULT, etc.) are retained.

+++ cpp/src/mavsdk/core/mavlink_component_metadata.cpp
@@ -96,7 +96,7 @@ void MavlinkComponentMetadata::receive_component_metadata(
                 _mavlink_components[message.compid].result = Result::Failed;
                 break;
         }
-        LogWarn("Requesting component metadata failed with {}", static_cast<int>(result));
+        LogWarn("Requesting component metadata failed with {}", result);
         on_all_types_completed(message.compid);
         return;
     }
