+++ cpp/src/mavsdk_server/src/plugins/gripper/gripper_service_impl.h
@@ -47,7 +47,7 @@ public:
     {
         switch (gripper_action) {
             default:
-                LogErr("Unknown gripper_action enum value: {}", static_cast<int>(gripper_action));
+                LogErr("Unknown gripper_action enum value: {}", gripper_action);
             // FALLTHROUGH
             case mavsdk::Gripper::GripperAction::Release:
                 return rpc::gripper::GRIPPER_ACTION_RELEASE;
@@ -61,7 +61,7 @@ public:
     {
         switch (gripper_action) {
             default:
-                LogErr("Unknown gripper_action enum value: {}", static_cast<int>(gripper_action));
+                LogErr("Unknown gripper_action enum value: {}", gripper_action);
             // FALLTHROUGH
             case rpc::gripper::GRIPPER_ACTION_RELEASE:
                 return mavsdk::Gripper::GripperAction::Release;
@@ -75,7 +75,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", static_cast<int>(result));
+                LogErr("Unknown result enum value: {}", result);
             // FALLTHROUGH
             case mavsdk::Gripper::Result::Unknown:
                 return rpc::gripper::GripperResult_Result_RESULT_UNKNOWN;
@@ -99,7 +99,7 @@ public:
     {
         switch (result) {
             default:
-                LogErr("Unknown result enum value: {}", static_cast<int>(result));
+                LogErr("Unknown result enum value: {}", result);
             // FALLTHROUGH
             case rpc::gripper::GripperResult_Result_RESULT_UNKNOWN:
                 return mavsdk::Gripper::Result::Unknown;
